--- lib/addedlocations.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class AddedLocations extends StatefulWidget {
-  const AddedLocations({Key? key}) : super(key: key);
-
-  @override
-  _AddedLocationsState createState() => _AddedLocationsState();
-}
-
-class _AddedLocationsState extends State<AddedLocations> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
