+++ lib/pages/onboarding/welcome_page.dart
