+++ lib/widgets/selector_card.dart
